Separated finders and style-stack-cache
